added unit 13

+++ app/Http/Controllers/Admin/UserController.php
@@ -15,6 +15,10 @@ use App\Http\Requests\{StoreUserRequest, UpdateUserRequest};
 
 class UserController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth:admin');
+    }
     /**
      * Display a listing of the resource.
      *
@@ -22,6 +26,8 @@ class UserController extends Controller
      */
     public function index()
     {
+        // dd(\Gate::allows('can-show-user'));
+        abort_unless(\Gate::allows('can-show-user'), 403);
         $users = User::orderBy('id', 'desc')->paginate();
         $title = 'Users Management';
         return view('admin.users.index', compact('users', 'title'));
@@ -34,6 +40,8 @@ class UserController extends Controller
      */
     public function create()
     {
+        abort_unless(\Gate::allows('can-create-user'), 403);
+
         $title = 'Add User';
         $status = UserStatus::toSelectArray();
         return view('admin.users.create', compact('title', 'status'));
@@ -47,6 +55,8 @@ class UserController extends Controller
      */
     public function store(StoreUserRequest $request)
     {
+        abort_unless(\Gate::allows('can-create-user'), 403);
+
         $user = User::create([
             'name' => $request['name'],
             'email' => $request['email'],
@@ -67,6 +77,7 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        abort_unless(\Gate::allows('can-show-user'), 403);
         return view('admin.users.show')->withUser($user)->withTitle('Show User');
     }
 
@@ -78,6 +89,8 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        abort_unless(\Gate::allows('can-edit-user'), 403);
+
         $status = UserStatus::toSelectArray();
         return view('admin.users.edit')->withUser($user)->withTitle('Edit User')->withStatus($status);
    
@@ -92,6 +105,8 @@ class UserController extends Controller
      */
     public function update(UpdateUserRequest $request, User $user)
     {
+        abort_unless(\Gate::allows('can-edit-user'), 403);
+
         $user->update($request->all());
         
         if (!$user->profile) {
@@ -118,6 +133,7 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        abort_unless(\Gate::allows('can-delete-user'), 403);
         $user->delete();
         return redirect()->route('admin.users.index');
     }
